rrr: convert reactor/fiber to module (rrr.fiber)
Header-only fiber.h (191 lines, just `this_fiber::*` inline wrappers
around Fiber::current_fiber/sleep) becomes the rrr.fiber module unit
at reactor/fiber.cpp. Sole consumer (rrr.hpp) switches from textual
include to `import rrr.fiber;`.
